added clone util to clone proposal and related objects
