+++ commercialoperator/components/proposals/utils.py
@@ -11,6 +11,7 @@ from commercialoperator.components.proposals.serializers import SaveProposalSeri
 from commercialoperator.components.main.models import Activity, Park, AccessType, Trail, Section, Zone
 import traceback
 import os
+from copy import deepcopy
 
 def create_data_from_form(schema, post_data, file_data, post_data_index=None,special_fields=[],assessor_data=False):
     data = {}
@@ -811,48 +812,6 @@ def save_assessor_data(instance,request,viewset):
         except:
             raise
 
-def clone_proposal_with_status_reset(proposal):
-    with transaction.atomic():
-        try:
-            proposal.customer_status = 'draft'
-            proposal.processing_status = 'draft'
-            proposal.assessor_data = {}
-            proposal.comment_data = {}
-
-            #proposal.id_check_status = 'not_checked'
-            #proposal.character_check_status = 'not_checked'
-            #proposal.compliance_check_status = 'not_checked'
-            #Sproposal.review_status = 'not_reviewed'
-
-            proposal.lodgement_number = ''
-            proposal.lodgement_sequence = 0
-            proposal.lodgement_date = None
-
-            proposal.assigned_officer = None
-            proposal.assigned_approver = None
-
-            proposal.approval = None
-
-            original_proposal_id = proposal.id
-
-            proposal.previous_proposal = proposal.objects.get(id=original_proposal_id)
-
-            proposal.id = None
-
-            #proposal.save(no_revision=True)
-            proposal.save()
-
-
-            # clone documents
-            for proposal_document in ProposalDocuments.objects.filter(proposal=original_proposal_id):
-                proposal_document.proposal = proposal
-                proposal_document.id = None
-                proposal_document.save()
-
-            return proposal
-        except:
-            raise
-
 def proposal_submit(proposal,request):
         with transaction.atomic():
             #import ipdb; ipdb.set_trace()
@@ -907,93 +866,6 @@ def proposal_submit(proposal,request):
                 raise ValidationError('You can\'t edit this proposal at this moment')
 
 
-def duplicate_object(self):
-    """
-    Duplicate a model instance, making copies of all foreign keys pointing to it.
-    There are 3 steps that need to occur in order:
+   
 
-        1.  Enumerate the related child objects and m2m relations, saving in lists/dicts
-        2.  Copy the parent object per django docs (doesn't copy relations)
-        3a. Copy the child objects, relating to the copied parent object
-        3b. Re-create the m2m relations on the copied parent object
 
-    """
-    related_objects_to_copy = []
-    relations_to_set = {}
-    # Iterate through all the fields in the parent object looking for related fields
-    for field in self._meta.get_fields():
-        if field.name in ['proposal', 'proposalrequest']:
-            print 'Continuing ...'
-            pass
-        elif field.one_to_many:
-            # One to many fields are backward relationships where many child objects are related to the
-            # parent (i.e. SelectedPhrases). Enumerate them and save a list so we can copy them after
-            # duplicating our parent object.
-            print('Found a one-to-many field: {}'.format(field.name))
-
-            # 'field' is a ManyToOneRel which is not iterable, we need to get the object attribute itself
-            related_object_manager = getattr(self, field.name)
-            related_objects = list(related_object_manager.all())
-            if related_objects:
-                print(' - {len(related_objects)} related objects to copy')
-                related_objects_to_copy += related_objects
-
-        elif field.many_to_one:
-            # In testing so far, these relationships are preserved when the parent object is copied,
-            # so they don't need to be copied separately.
-            print('Found a many-to-one field: {}'.format(field.name))
-
-        elif field.many_to_many:
-            # Many to many fields are relationships where many parent objects can be related to many
-            # child objects. Because of this the child objects don't need to be copied when we copy
-            # the parent, we just need to re-create the relationship to them on the copied parent.
-            print('Found a many-to-many field: {}'.format(field.name))
-            related_object_manager = getattr(self, field.name)
-            relations = list(related_object_manager.all())
-            if relations:
-                print(' - {} relations to set'.format(len(relations)))
-                relations_to_set[field.name] = relations
-
-    # Duplicate the parent object
-    self.pk = None
-    self.lodgement_number = ''
-    self.save()
-    print('Copied parent object {}'.format(str(self)))
-
-    # Copy the one-to-many child objects and relate them to the copied parent
-    for related_object in related_objects_to_copy:
-        # Iterate through the fields in the related object to find the one that relates to the
-        # parent model (I feel like there might be an easier way to get at this).
-        for related_object_field in related_object._meta.fields:
-            if related_object_field.related_model == self.__class__:
-                # If the related_model on this field matches the parent object's class, perform the
-                # copy of the child object and set this field to the parent object, creating the
-                # new child -> parent relationship.
-                related_object.pk = None
-                #if related_object_field.name=='approvals':
-                #    related_object.lodgement_number = None
-                if isinstance(related_object, Approval):
-                    related_object.lodgement_number = ''
-
-                setattr(related_object, related_object_field.name, self)
-                print related_object_field
-                try:
-                    related_object.save()
-                except:
-                    import ipdb; ipdb.set_trace()
-
-                text = str(related_object)
-                text = (text[:40] + '..') if len(text) > 40 else text
-                print('|- Copied child object {}'.format(text))
-
-    # Set the many-to-many relations on the copied parent
-    for field_name, relations in relations_to_set.items():
-        # Get the field by name and set the relations, creating the new relationships
-        field = getattr(self, field_name)
-        field.set(relations)
-        text_relations = []
-        for relation in relations:
-            text_relations.append(str(relation))
-        print('|- Set {} many-to-many relations on {} {}'.format(len(relations), field_name, text_relations))
-
-    return self
